chore: Complete OpenSpec validation and code quality improvements
Complete Phase 6 (Testing & Validation) and Phase 7 (Documentation) of the
fix-gl-renderer-bcf-serialization OpenSpec change.

Code Quality Fixes:
- Fix clippy warnings: remove unused enumerate(), unnecessary casts
- Add `#[allow(dead_code)]` to bcf_cpu_tracer (WIP module)
- Add safety documentation to all unsafe functions in gl_tracer and gpu_tracer
- Fix empty lines after doc comments

Documentation Enhancements:
- Expand GL tracer module docs with detailed BCF serialization approach
- Document GPU upload strategy (1D-like 2D texture, R8UI format)
- Add inline comments explaining BCF node format and coordinate systems
- Update OpenSpec tasks.md to mark all phases complete

Testing Validation:
- Verified renderer runs successfully with BCF serialization (21 bytes)
- Confirmed shader compilation and GPU buffer upload working
- Validated clippy passes with -D warnings
- Confirmed cargo fmt compliance

All phases (1-7) of the BCF GL renderer fix are now complete.

+++ crates/cube/src/traversal/mod.rs
@@ -66,7 +66,7 @@ pub fn traverse_octree(grid: &NeighborGrid, visitor: TraversalVisitor, max_depth
         let grid_z = (octant_pos.z + 3) / 2;
         let grid_idx = NeighborGrid::xyz_to_index(grid_x, grid_y, grid_z);
 
-        let view = NeighborView::new(grid, grid_idx as usize);
+        let view = NeighborView::new(grid, grid_idx);
         let coord = CubeCoord::new(octant_pos, max_depth);
         traverse_recursive(view, coord, visitor, false);
     }
@@ -108,7 +108,7 @@ fn traverse_recursive(
                     let grid_z = (octant_pos.z + 3) / 2;
                     let child_grid_idx = NeighborGrid::xyz_to_index(grid_x, grid_y, grid_z);
 
-                    let child_view = NeighborView::new(&child_grid, child_grid_idx as usize);
+                    let child_view = NeighborView::new(&child_grid, child_grid_idx);
                     let child_coord = coord.child(octant_idx);
                     traverse_recursive(child_view, child_coord, visitor, true);
                 }
@@ -127,7 +127,7 @@ fn traverse_recursive(
                 let grid_z = (octant_pos.z + 3) / 2;
                 let child_grid_idx = NeighborGrid::xyz_to_index(grid_x, grid_y, grid_z);
 
-                let child_view = NeighborView::new(&child_grid, child_grid_idx as usize);
+                let child_view = NeighborView::new(&child_grid, child_grid_idx);
                 let child_coord = coord.child(octant_idx);
                 traverse_recursive(child_view, child_coord, visitor, false);
             }
